Pass 1: Determine method to obfuscate
Previous version had a dependency problem (annotation required before
the visit, but obtained during it). Now the parsing is done three
times. In the first pass, the method with the @Obfuscate annotation are
saved in an ArrayList.
Also, updated the import in order to not rely on jdk internal status