support for unicode classes and methods
Something that I never encountered in real scenarios, but java actually
supports unicode identifiers. Tried this and java2c was not able to
transform correctly, so I fixed this bug.

However, I really hope that nobody and really nobody would ever use a
non-ASCII identifier, for the sake of comprehension.