Fix exceptions under new antidebug system
Exceptions detection and resolution used some hardcoded values (given
that they require the first position of the stack) and this effectively
bypassed the push/pop implementation sending data to the child vm and
instead used a local, empty, placeholder stack.

Moreover, exceptions resolution instantly returned, bypassing the
method cleanup logic. This logic is essential in the self debugging
implementation, given that the child vm uses a stacks to track each
methods status. By skipping this logic when returning, the child vm
was never signalled that the method returned.