Fix SIGFPE in CSPRNG longjump
In the antidebugging with self debugging, both parent and child choose a
number that is later used to handle the reseed. Specifically, the reseed
happens when the generated number mod the chosen constant is zero.
However, I never checked whether the chosen constant was itself zero,
and this lead to several SIGFPE.