Fixed race condition
This was a tough one to find. After sending the kill signal the parent
was never waiting the child to terminate. In very fast sequential
execution it could happen that the parent was invoking a new child while
the old one was still due to be terminated. This could create an awesome
situation where the child to be killed was still attached to the parent,
and the newborn child was trying to attach to the parent, failing.

With this patch, the parent waits until the child is effectively dead
before proceeding with the execution.