Better error handling for NativeCompiler class
Previous error handling was done C-style, with a String that was null if
no error happened or contained the error message otherwise.
No annotation whatsoever to enforce checking the return value, so
NATURALLY this was ignored in the main application.

New implementation throws an exception in case of error so it must be
handled properly in any case. Additionally, the InterruptedException
that was previously ignored now must be catched (because interrupting
the compilation is the same as compilation error, so the application
expects a correct compilation but instead there is a failure).